add testCollectFishBonesOnMiddleLane
Add a test that compares the number of coins displayed at the end against the number of fishbones the cat collect from the middle lane. This is a "good-enough" test that covers a good amount of cases. It doesn't cover the cases where the cat reaches 150m+, when origin resets occurs for most fishbones; this needs to be handled